+++ PLC_esp8266/main/Maintenance/ServiceModeHandler.cpp
@@ -1,17 +1,15 @@
 #include "Maintenance/ServiceModeHandler.h"
 #include "Display/Common.h"
-#include "Display/LogsList.h"
 #include "Display/display.h"
 #include "buttons.h"
 #include "esp_err.h"
 #include "esp_log.h"
 #include "settings.h"
-#include "smartconfig_service.h"
 #include "sys_gpio.h"
 #include <stdlib.h>
 #include <string.h>
 
-static const char *TAG_ServiceModeHandler = "service_mode";
+static const char *TAG_ServiceModeHandler = "ServiceMode";
 
 #define EXPECTED_BUTTONS                                                                           \
     (BUTTON_UP_IO_CLOSE | BUTTON_UP_IO_OPEN | BUTTON_DOWN_IO_CLOSE | BUTTON_DOWN_IO_OPEN           \
@@ -135,76 +133,4 @@ void ServiceModeHandler::Execute(EventGroupHandle_t gpio_events, Mode mode) {
             ESP_LOGE(TAG_ServiceModeHandler, "Mode::sm_ResetToDefault not implemented");
             break;
     }
-}
-
-void ServiceModeHandler::SmartConfig(EventGroupHandle_t gpio_events) {
-    ESP_LOGI(TAG_ServiceModeHandler, "exec SmartConfig");
-    uint8_t *fb;
-    LogsList logs_list("SmartConfig");
-
-    bool runned = true;
-    start_smartconfig();
-    while (runned) {
-        switch (smartconfig_status()) {
-            case scs_Start:
-                logs_list.Append("Start");
-                break;
-
-            case scs_Started:
-                logs_list.Append("Started");
-                break;
-
-            case scs_Disconnected:
-                logs_list.Append("Disconnected 0123456");
-                break;
-
-            case scs_GotIP:
-                logs_list.Append("Got IP");
-                break;
-
-            case scs_ScanDone:
-                logs_list.Append("Scanning");
-                break;
-
-            case scs_FoundChannel:
-                logs_list.Append("Found channel");
-                break;
-
-            case scs_GotCreds:
-                logs_list.Append("Got credentials");
-                break;
-
-            case scs_Completed:
-                logs_list.Append("Completed");
-                runned = false;
-                break;
-
-            case scs_Error:
-                logs_list.Append("Error");
-                runned = false;
-                break;
-        }
-
-        fb = begin_render();
-        logs_list.Render(fb);
-        end_render(fb);
-    }
-
-    uint8_t x = 1;
-    uint8_t y = 1;
-    uint8_t height = get_text_f6X12_height();
-
-    const int show_logs_time_ms = 3000;
-    xEventGroupWaitBits(gpio_events,
-                        EXPECTED_BUTTONS,
-                        true,
-                        false,
-                        show_logs_time_ms / portTICK_PERIOD_MS);
-
-    fb = begin_render();
-    ESP_ERROR_CHECK(draw_text_f6X12(fb, x, y + height * 1, "SC completed!") <= 0);
-    ESP_ERROR_CHECK(draw_text_f6X12(fb, x, y + height * 2, "Press SELECT to exit") <= 0);
-    end_render(fb);
-
-    xEventGroupWaitBits(gpio_events, EXPECTED_BUTTONS, true, false, portMAX_DELAY);
 }
