+++ PLC_esp8266/main/Maintenance/ServiceModeHandler.h
@@ -12,6 +12,10 @@ extern "C" {
 }
 #endif
 
+#define EXPECTED_BUTTONS                                                                           \
+    (BUTTON_UP_IO_CLOSE | BUTTON_UP_IO_OPEN | BUTTON_DOWN_IO_CLOSE | BUTTON_DOWN_IO_OPEN           \
+     | BUTTON_SELECT_IO_CLOSE | BUTTON_SELECT_IO_OPEN)
+
 class ServiceModeHandler {
   public:
     enum Mode { sm_SmartConfig, sm_BackupLogic, sm_RestoreLogic, sm_ResetToDefault };
